Of course rp.available_mods should be accessible...
rp.available_mods is now a "shadow" copy of the original one. It contains all current values, but is read-only in a way. Mods should not mess around with other mods' entries and I don't really want to make outdated copies as some mods might depend upon the loading status of other mods.