Fixed double-require bug and set version to 1400.
The RP naming scheme has changed and this time I hope it's final. The last two digits are the current build (in this case, 00). Odd builds are used during development and therefore might be incomplete and/or completely broken. Even builds are usually released. The other digits represent the current release, i.e. "1435" would be "Alpha Release 14, build 35".